minor refactor, add ability to turn off xml/json-index output
