+++ src/shared/ogcapi-records/src/main/java/org/geonetwork/ogcapi/service/formatter/FormatterApiMessageWriter.java
@@ -3,7 +3,7 @@
  * This code is licensed under the GPL 2.0 license,
  * available at the root application directory.
  */
-package org.geonetwork.ogcapi.configuration;
+package org.geonetwork.ogcapi.service.formatter;
 
 import java.io.IOException;
 import java.util.HashMap;
